fixed type problem with threshold
git-svn-id: svn://orca.physics.unc.edu/Orca@3209 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Experiments/BaseClasses/ORDetectorSegment.h
@@ -71,7 +71,7 @@
 - (void) setSegmentError:(BOOL)state;
 - (void) clearSegmentError;
 - (void) setSegmentError;
-- (short) threshold;
+- (unsigned long) threshold;
 - (void) setGain:(id)aValue;
 - (void) setThreshold:(id)aValue;
 - (float) totalCounts;
